+++ packages/app/src/components/VisualizationOptions/Options/SelectBaseOption.js
@@ -82,9 +82,7 @@ SelectBaseOption.propTypes = {
 }
 
 const mapStateToProps = (state, ownProps) => ({
-    value:
-        sGetUiOptions(state)[ownProps.option.name] ||
-        ownProps.option.defaultValue,
+    value: sGetUiOptions(state)[ownProps.option.name],
 })
 
 const mapDispatchToProps = (dispatch, ownProps) => ({